Move CFI analysis out of MC layer
