+++ llvm/lib/MC/CMakeLists.txt
@@ -87,4 +87,3 @@ add_llvm_component_library(LLVMMC
 
 add_subdirectory(MCParser)
 add_subdirectory(MCDisassembler)
-add_subdirectory(CFIAnalysis)
