+++ llvm/lib/MCCFIAnalysis/CMakeLists.txt
@@ -3,6 +3,9 @@ add_llvm_component_library(LLVMMCCFIAnalysis
   CFIAnalysisMCStreamer.cpp
   CFIAnalysis.cpp
 
+  ADDITIONAL_HEADER_DIRS
+  ${LLVM_MAIN_INCLUDE_DIR}/llvm/MCA
+
   LINK_COMPONENTS
   MC
   Support
