+++ llvm/tools/llvm-mc/llvm-mc.cpp
@@ -14,7 +14,7 @@
 #include "Disassembler.h"
 #include "llvm/MC/MCAsmBackend.h"
 #include "llvm/MC/MCAsmInfo.h"
-#include "llvm/MC/MCCFIAnalysis/CFIAnalysisMCStreamer.h"
+#include "llvm/MCCFIAnalysis/CFIAnalysisMCStreamer.h"
 #include "llvm/MC/MCCodeEmitter.h"
 #include "llvm/MC/MCContext.h"
 #include "llvm/MC/MCInstPrinter.h"
